requirements.txt
transformers>=4.18.0
tqdm>=4.64.0
torchvision>=0.12.0
torch>=1.11.0
opencv-python>=4.5.4.60
imageio>=2.19.2
ffmpeg-python>=0.2.0
Pillow>=9.1.1
wget>=3.2
omegaconf>=2.1.2
kornia>=0.6.4
regex>=2022.4.24
numpy>=1.21.4
taming-transformers>=0.0.01
pandas>=1.5.1
timm>=0.6.11
gdown>=4.6.0
mega.py>=1.0.8
accelerate>=0.15.0
tensorboard>=2.11.0
diffusers>=0.11.1
moviepy>=1.0.3
